move editor panel styles to its own file
